+++ mqmetric/log.go
@@ -46,7 +46,8 @@ func logInfo(format string, v ...interface{}) {
 	}
 }
 
-// Errors should be reported always
+// Errors should be reported always. Also use this for what you might
+// think of as warnings.
 func logError(format string, v ...interface{}) {
 	if logger != nil && logger.Error != nil {
 		logger.Error(format, v...)